pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anki-gen"
version = "0.1.0"
description = "Generate Anki flashcards from Markdown notes using an LLM"
requires-python = ">=3.10"
dependencies = [
    "genanki>=0.13.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "markdown-it-py>=3.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
]

[project.scripts]
anki-gen = "anki_gen.cli:main"
anki-gen-mcp = "anki_gen.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["anki_gen"]
